backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "REST API for minimal TODO Application"
authors = ["lestherll"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn-worker = "^0.0.1"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.18.0"}
python-multipart = "^0.0.5"
fastapi-login = "^1.7.3"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
